+++ BTCPayServer/Plugins/Shopify/ShopifyPlugin.cs
@@ -14,7 +14,8 @@ namespace BTCPayServer.Plugins.Shopify
 
         public override void Execute(IServiceCollection applicationBuilder)
         {
-            applicationBuilder.AddSingleton<IHostedService, ShopifyOrderMarkerHostedService>();
+            applicationBuilder.AddSingleton<ShopifyService>();
+            applicationBuilder.AddSingleton<IHostedService, ShopifyService>(provider => provider.GetRequiredService<ShopifyService>());
             applicationBuilder.AddSingleton<IUIExtension>(new UIExtension("Shopify/NavExtension", "header-nav"));
             base.Execute(applicationBuilder);
         }
